refactor(fspy): extract Unix client crate
