+++ Cargo.toml
@@ -77,6 +77,7 @@ flate2 = "1.0.35"
 fspy = { path = "crates/fspy" }
 fspy_benchmark_launcher = { path = "crates/fspy_benchmark_launcher", artifact = "bin" }
 fspy_benchmark_target = { path = "crates/fspy_benchmark_target", artifact = "bin" }
+fspy_client_unix = { path = "crates/fspy_client_unix" }
 fspy_detours_sys = { path = "crates/fspy_detours_sys" }
 fspy_preload_unix = { path = "crates/fspy_preload_unix", artifact = "cdylib", target = "target" }
 fspy_preload_windows = { path = "crates/fspy_preload_windows", artifact = "cdylib", target = "target" }
@@ -224,6 +225,11 @@ panic = "abort" # Let it crash and force ourselves to write safe Rust.
 # `strip` must be restated: a package override resets an inherited `strip`
 # to "debuginfo" when the field is omitted (Cargo quirk in profile merging).
 [profile.release.package.fspy_preload_unix]
+opt-level = 3
+codegen-units = 1
+strip = "symbols"
+
+[profile.release.package.fspy_client_unix]
 opt-level = 3
 codegen-units = 1
 strip = "symbols"
