+++ crates/fspy_client_unix/.clippy.toml
@@ -1,0 +1,1 @@
+../../.non-vite.clippy.toml
