+++ crates/fspy_preload_unix/Cargo.toml
@@ -8,20 +8,14 @@ publish = false
 crate-type = ["cdylib"]
 
 [target.'cfg(unix)'.dependencies]
-allocator-api2 = { workspace = true, features = ["alloc"] }
-anyhow = { workspace = true }
-wincode = { workspace = true }
-bstr = { workspace = true, features = ["alloc", "std"] }
 ctor = { workspace = true }
+fspy_client_unix = { workspace = true }
 fspy_shared = { workspace = true }
 fspy_shared_unix = { workspace = true }
 libc = { workspace = true }
 nix = { workspace = true, features = ["signal", "fs", "socket", "mman", "time"] }
 sigsafe = { workspace = true }
 sigsafe_alloc = { workspace = true }
-
-[target.'cfg(target_os = "linux")'.dependencies]
-itoa = { workspace = true }
 
 [build-dependencies]
 artifact_profile = { workspace = true }
