+++ crates/fspy_preload_unix/src/client/mod.rs
@@ -1,121 +1,7 @@
-pub mod convert;
-pub mod raw_exec;
-
-use std::{
-    cell::Cell, ffi::OsStr, fmt::Debug, num::NonZeroUsize, os::unix::ffi::OsStrExt as _,
-    path::Path, sync::OnceLock,
-};
+use std::{cell::Cell, sync::OnceLock};
 
 use convert::{ToAbsolutePath, ToAccessMode};
-use fspy_shared::ipc::{PathAccess, channel::Sender};
-use fspy_shared_unix::{
-    exec::ExecResolveConfig,
-    payload::EncodedPayload,
-    spawn::{PreExec, handle_exec},
-};
-use raw_exec::RawExec;
-use wincode::Serialize as _;
-
-pub struct Client {
-    encoded_payload: EncodedPayload,
-    ipc_sender: Option<Sender>,
-}
-
-// SAFETY: Client fields are only mutated during initialization in the ctor; after that, all access is read-only
-#[cfg(target_os = "macos")]
-unsafe impl Sync for Client {}
-// SAFETY: Client is only sent once during initialization; after that it lives in a static OnceLock
-#[cfg(target_os = "macos")]
-unsafe impl Send for Client {}
-
-impl Debug for Client {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Client").finish()
-    }
-}
-
-impl Client {
-    #[expect(
-        clippy::print_stderr,
-        reason = "preload library intentionally uses stderr for error reporting"
-    )]
-    #[cfg(not(test))]
-    fn from_env(envs: impl Iterator<Item = sigsafe::env::Entry>) -> Self {
-        use fspy_shared_unix::payload::decode_payload_from_env;
-
-        let encoded_payload = decode_payload_from_env(envs).unwrap();
-
-        let ipc_sender = match encoded_payload.payload.ipc_channel_conf.sender() {
-            Ok(sender) => Some(sender),
-            Err(err) => {
-                // this can happen if the process is started after the root target process has exited.
-                // By that time the channel would have been closed in the receiver side.
-                // In this case we just leave a message and skip sending any path accesses.
-                eprintln!("fspy: failed to create ipc sender: {err}");
-                None
-            }
-        };
-
-        Self { encoded_payload, ipc_sender }
-    }
-
-    fn send(&self, mode: fspy_shared::ipc::AccessMode, path: &Path) -> anyhow::Result<()> {
-        let Some(ipc_sender) = &self.ipc_sender else {
-            // ipc channel not available, skip sending
-            return Ok(());
-        };
-        let path_bytes = path.as_os_str().as_bytes();
-        if path_bytes.starts_with(b"/dev/")
-            || (cfg!(target_os = "linux")
-                && (path_bytes.starts_with(b"/proc/") || path_bytes.starts_with(b"/sys/")))
-        {
-            return Ok(());
-        }
-        let path_access = PathAccess { mode, path: path.into() };
-        let serialized_size = usize::try_from(PathAccess::serialized_size(&path_access)?)
-            .expect("serialized size exceeds usize");
-
-        let frame_size = NonZeroUsize::new(serialized_size)
-            .expect("fspy: encoded PathAccess should never be empty");
-
-        let mut frame = ipc_sender
-            .claim_frame(frame_size)
-            .expect("fspy: failed to claim frame in shared memory");
-        let mut writer: &mut [u8] = &mut frame;
-        PathAccess::serialize_into(&mut writer, &path_access)?;
-        assert_eq!(writer.len(), 0);
-
-        Ok(())
-    }
-
-    pub unsafe fn handle_exec<R>(
-        &self,
-        config: ExecResolveConfig,
-        raw_exec: RawExec,
-        f: impl FnOnce(RawExec, Option<PreExec>) -> nix::Result<R>,
-    ) -> nix::Result<R> {
-        // SAFETY: raw_exec contains valid pointers to C strings and null-terminated arrays, as provided by the caller
-        let mut exec = unsafe { raw_exec.to_exec() };
-        let pre_exec = handle_exec(&mut exec, config, &self.encoded_payload, |mode, path| {
-            self.send(mode, path).unwrap();
-        })?;
-        RawExec::from_exec(exec, |raw_command| f(raw_command, pre_exec))
-    }
-
-    pub unsafe fn try_handle_open(
-        &self,
-        path: impl ToAbsolutePath,
-        mode: impl ToAccessMode,
-    ) -> anyhow::Result<()> {
-        // SAFETY: mode contains a valid pointer (if ModeStr) or a plain value, as provided by the caller
-        let mode = unsafe { mode.to_access_mode() };
-        let arena = sigsafe_alloc::arena();
-        let Some(abs_path) = path.to_absolute_path(&arena)? else {
-            return Ok(());
-        };
-        self.send(mode, Path::new(OsStr::from_bytes(abs_path.as_bytes())))
-    }
-}
+pub use fspy_client_unix::{Client, convert, raw_exec};
 
 static CLIENT: OnceLock<Client> = OnceLock::new();
 
@@ -146,7 +32,8 @@ pub unsafe fn handle_open(path: impl ToAbsolutePath, mode: impl ToAccessMode) {
         let _reset = ResetHandling(handling);
 
         if let Some(client) = global_client() {
-            // SAFETY: path and mode contain valid pointers/values forwarded from the interposed function's caller
+            // SAFETY: path and mode contain valid pointers/values forwarded
+            // from the interposed function's caller.
             unsafe { client.try_handle_open(path, mode) }.unwrap();
         }
     });
